Renamed controller/router for simplicity, modified server.js to reflect changes, added my own .env. DB is running on my end.
Controller, model, db, and routers should be worked on ASAP!
