+++ server/.env.example
@@ -15,3 +15,4 @@ VM_PASS=
 DB_USER=
 DB_PASS=
 DB_NAME=
+DB_PORT=
